==> 1.0_Compile.Rmd <==
# Set Up
```{r setup, include = FALSE}
# librarys
library(tidyverse)
library(dplyr)
library(lubridate)
library(chron)
library(lunar)
library(ggplot2)
```

# Data Prep
After downloading the data from arcCollector, you end up four individual .csv docs for each survey layer; 

  - Meta
  - Extractive Use 
  - Non-Extractive Use
  - Avidity

1. Read in documents, including `na = c("", " ")` to make all empty cells consistent

```{r  read data}
# Note this data has been manually compiled across 2 feature classes after updating the data
meta <- read_csv("./data/RAW/Meta.csv", na =c("", " ", "na", "NA", "-"))
ex <- read_csv("./data/RAW/Extractive.csv", na =c("", " ", "na", "NA", "-"))
nex <- read_csv("./data/RAW/NonExtractive.csv", na =c("", " ", "na", "NA", "-"))
avid <- read_csv("./data/RAW/Avidity.csv", na =c("", " ", "na", "NA", "-"))
```

Do not expect any of your data sheets to have the same number of observations.

  - Meta data will contain one observation per person approached so wil include refusals.
  
  - Extractive and non-extractive use data will contain an observation per use which may
  be more than one per person therefore could have more observations than the meta data.
  Alternatively it may have less observations than the meta data if most people doing the survey are
  participating in the alternative use type (i.e. most people are doing extractive and few are doing
  non-extractive)
  
  - Avidity data will contain an observation per completed survey, so you would expect there to be
  less observations in the avidity data than the meta data as it does not include refusals. 

For each individual dataset: 
  2. Check no missing GlobalIDs in Meta or GUIDs in Use or Avidity data
  3. Give ID variables intuitive names
  5 Add activity type for Use surveys

```{r  data prep}
table(is.na(meta$GlobalID)) # all false

meta %<>% # meta data
  rename(mGlobID = GlobalID,
         mLong = x,
         mLat = y) %>%
  dplyr::select(-c(created_user, last_edited_user, last_edited_date, "GUID for meta"))

table(is.na(ex$exGUID)) # all false

ex %<>% # extractive data
  rename(exGlobID = GlobalID,
         exLong = x,
         exLat = y) %>%
  dplyr::select(-c(created_user, last_edited_user, last_edited_date)) %>%
  mutate(exActivityType = "Extractive")

table(is.na(nex$nexGUID)) # all false

nex %<>% # non-extractive data
  rename(nexGlobID = GlobalID,
         nexLong = x,
         nexLat = y) %>%
  dplyr::select(-c(created_user, last_edited_user, last_edited_date)) %>%
  mutate(nexActivityType = "NonExtractive")

table(is.na(avid$aGUID)) # all false

avid <- avid %>% # avidity data
  dplyr::select(-c(created_user, last_edited_user, last_edited_date, GlobalID))
```

Join the four data sets, linking them by the Meta GlobalID (primary key) and GUID (foreign key) of the other data sheets (ie. every Meta GlobalID should match the GUID in the other data sheets). 

meta (GlobalID) -> ex/nex/avid GUID

Use a `full_join` to connect individual data sheets. Using a inner join will remove refusals which are needed for response rate calculations. 

```{r  join}
dat1 <- full_join(meta, ex, by = c("mGlobID" = "exGUID"), keep = T) # joining meta to extractive
dat2 <- full_join(dat1, nex, by = c("mGlobID" = "nexGUID"), keep = T) # joining non-extractive
dat <- full_join(dat2, avid, by = c("mGlobID" = "aGUID"), keep = T) # joining avidity
```

```{r save}
write.csv(dat, './data/RAW/marm_RAW.csv')  # full raw un-cleaned dataset
```
