==> 1.1_Clean.Rmd <==
# Set Up
```{r setup, include = FALSE}
# librarys
library(tidyverse)
library(dplyr)
library(lubridate)
library(chron)
library(lunar)
library(ggplot2)

# functions
source("./functions/cleaningFunc.R") # spatial functions

# data
dat <- read_csv("./data/RAW/RAW.csv")
```

## Rename
Rename questions to make them R friendly
```{r select and filter}
dat  <- dat %>% 
dplyr::rename(ID = "...1",
              PersonID = "OBJECTID.x",
              Interviewer = "# Interviewer",
              Site = "# Site",
              Screen18 = "Screen: 18+",
              PrevInter = "Screen: Have you done this survey recently?", 
              Agreement = "Agreement to participate",
              Resident = "Are you/have you been a resident, got a holiday home here or been here over 3 months?",
              
              nDaysInArea = "If here for less than 3 months, how many DAYS will you be in the area in total on this trip?",
              BoatAccess = "While you are on your trip, do you have access to a motorised boat/jetski (private/rental)?",
              nBoatDays = "If YES, how many DAYS out of your trip will you be going out on a boat in total?",
              nShoreDays = "How many DAYS of your trip will you be doing coastal shore based activities in total?",

              nTimesLast12m = "Excluding this trip, how many times have you been to the area in the last 12 month?",
              SurveyDate = "created_date.x",
              exSiteType = "# Is this a boat based or shore based activity?",
              exRecall = "If RECALL, what date did they go fishing?",
              FishingType = "What type of fishing did you do?",
              BaitLure = "Did you use bait or lure?",
              exStart = "What time did you START fishing at this site?",
              exStop = "What time did you STOP fishing at this site?",
              MaxHook = "What was your maximum hook depth (m) at this site?",
              KeptUndam = "Number of fish kept undamaged at this site?",
              RelUndam = "How many fish did you release at this site?",
              nDP = "How many shark bite offs did you get at this site?",
              Species = "What species of fish did you keep?", 
              exWhyLeave = "Why did you leave this site?",
              exComments = "Comments.x",
              nexSiteType = "# Is this a boat-based or shore-based activity?",
              nexRecall = "If RECALL, what date did they participant do the activity?",
              Activity = "What activity did you do?",
              nexStart = "What time did you START the activity?",
              nexStop = "What time did you STOP the activity?",
              DiveMaxDepth = "If DIVING, what was your maximum depth (m)?",
              WhyChoose = "Why did you CHOOSE this site?",
              nexWhyLeave = "Why did you LEAVE this site?",
              nexTimes12m = "How many DAYS have you done this activity in the last 12 months?",
              nexYrs = "How many YEARS have you been doing this activity?",
              nexComments = "Comments.y",
              exTimes12m = "How many days have you fished in the last 12 months?",
              exYrs = "How many years have you been fishing for?",
              DiveCert = "Are you a scuba diver? If so, what is your certification?",
              nDives = "If SCUBA DIVER, how many dives have you done?",
              YrBorn = "What year were you born?",
              Postcode = "What is your postcode?",
              Accom = "If VISITOR, where are you staying?",
              Sex = "# Female or Male",
              nFemales = "# Number of females in party (excluding participant)",
              nMales = "# Number of males in party (excluding participant)",
              nGirls = "# Number of girls in party",
              nBoys = "# Number of boys in party",
              BoatLength = "# Boat Length (m)",
              BoatID = "# Boat number",
              aComments = "Comments",
              exDetract = "Did anything detract from your experience at this site?.x",
              nexDetract = "Did anything detract from your experience at this site?.y",
              Native = "Are you of Aboriginal or Torres Strait Islander origin?",
              TO = "Are you a Traditional Owner of the region?",
              exCustom = "If of Aboriginal or Torres Strait Islander origin, was this activity undertaken for a customary purpose?",
              nexCustom = "If of Aboriginal or Torres Strait Islander origin, was this activity undertaken from a customary purpose?")
```

## Temporal variables
  - The date column is automated so there should be no NA's unless the agreement is no. 
  - Adhere start and stop times for extractive and non-extractive activities into one column,
    there should only be blanks where agreement is no. 
  - Due to recall trips need to make a trip date column which may be different to survey date.
  - Any data out with field trip dates is a dummy/test survey
  
```{r}
dat <- dat %>% 
  mutate(SurveyDate = ifelse(!is.na(SurveyDate), SurveyDate,
                       ifelse(is.na(SurveyDate), created_date.x, 
                              ifelse(is.na(created_date.y), created_date.x.x, created_date.y.y)))) %>% 
  # mutate(SurveyDate = sub(" .*", "", SurveyDate)) %>% 
  mutate(numYear = as.numeric(format(as.Date(SurveyDate, format = "%m/%d/%Y"),"%Y"))) %>%  # year (num)
  mutate(facYear = as.factor(numYear)) %>% # year (chr)
  mutate(TripDate = ifelse(!is.na(exRecall), exRecall, nexRecall)) %>% 
  mutate(TripDate = ifelse(is.na(TripDate), SurveyDate, TripDate)) %>%  
  mutate(TripDate = sub(" .*", "", TripDate)) %>% 
  mutate(TripDate = as.Date(TripDate, "%m/%d/%Y")) %>% 
  mutate(TripMonth = as.factor(format(TripDate,"%B"))) %>%  # month (chr)
  mutate(TripJulianDay = as.numeric(julian(TripDate))) %>% # julian date (num)
  mutate(binRecall = as.factor(ifelse(TripDate == SurveyDate, "0", "1"))) %>% # binary recall (fac)
  dplyr::select(-c(exRecall, nexRecall, created_date.y, created_date.x.x, created_date.y.y))


# Start (chr), Stop (chr), MedianTime (chr), decMedianTime (num), decDuration (num)

dat <- dat %>%
  mutate(Start = substr(as.character(ifelse(!is.na(exStart), as.character(exStart),
                                     as.character(nexStart))), 1, 5))  %>%  # start time
   mutate(Stop = substr(as.character(ifelse(!is.na(exStop), as.character(exStop),
                                     as.character(nexStop))), 1, 5)) %>%    # start time
  mutate(temp_Start = parse_date_time(Start, c("HM"))) %>%   # making temp_start POSIX for calcs
  mutate(temp_Stop = parse_date_time(Stop, c("HM"))) %>% # making temp_stop POSIX for calcs
  rowwise() %>%
  mutate(MedianTime = substr(median(c(temp_Start, temp_Stop)), 12, 16)) %>%  # median time (chr)
  ungroup() %>% 
  mutate(decMedianTime = as.numeric(round(hhmm2dec(MedianTime), 
                                          digits = 2))) %>% #decimal median time (num)
  mutate(decDuration = as.numeric(round(difftime(
    temp_Stop, temp_Start, units = "hours"), digits = 2))) # decimal duration of activity (num)
```


## IDs
  -TripNum: all activities on one boat launch (individual date) are collectively a trip, only       applicable is data from recalled trips was collected (if TripNum >1 then binRecall should        equal 1). TripNum is chronological. 
  -SiteNum: the chronological order a site was visited within one trip
  
  * Mutate Person ID to reflect that people with same boat ID should have the same associated person ID. 

```{r IDs}
dat <- dat %>%
  group_by(PersonID) %>%
  arrange(Start) %>% 
  mutate(TripNum = as.numeric(as.factor(TripDate))) %>% # number trips done by same person across different days
  mutate(SiteNum = sequence(rle(TripNum)$lengths)) %>% # chronological order of which a participant has conducted uses in a single trip
  arrange(ID) %>%
  dplyr::ungroup()

# survey ID: is the ID of an individual survey
# person ID: the id of a person via there boat number
# trip id: an id per individual day trip per person 

dat <- dat %>% 
  mutate(BoatID = toupper(BoatID)) %>% 
  group_by(BoatID) %>% 
  mutate(PersonID = ifelse(anyDuplicated(BoatID) & !is.na(BoatID), min(PersonID), PersonID)) %>% 
  ungroup() %>% 
  group_by(PersonID, TripDate) %>% 
  mutate(TripID = cur_group_id()) %>% 
  ungroup()
```

## Meta cleaning
```{r meta}
dat <- dat %>% 
  mutate(nBoatDays = as.numeric(ifelse(nBoatDays %in% "All", nBoatDays, nBoatDays))) %>% 
  mutate(nShoreDays = as.numeric(ifelse(nShoreDays %in% "All", nDaysInArea, nShoreDays))) %>% 
  mutate_each_(~factor(.), c("Interviewer", "Screen18", "PrevInter", "Agreement",
                                  "BoatAccess", "Resident", "facYear"))
```

## Use cleaning
```{r use surveys}
dat <- dat %>% 
  mutate(SiteType = as.factor(ifelse(!is.na(exSiteType), exSiteType, nexSiteType))) %>% 
  mutate(ActivityType = as.factor(ifelse(!is.na(exActivityType), exActivityType, nexActivityType))) %>% 
  mutate(WhyLeave = as.factor(ifelse(!is.na(exWhyLeave), exWhyLeave, nexWhyLeave))) %>% 
  mutate(Detract = as.factor(ifelse(!is.na(exDetract), exDetract, nexDetract))) %>% 
  mutate(UseLat = ifelse(!is.na(exLat), exLat, nexLat)) %>% 
  mutate(UseLong = ifelse(!is.na(exLong), exLong, nexLong)) %>% 
  mutate(Activity = as.factor(ifelse(ActivityType == "Extractive", "Fishing", Activity))) %>% 
  mutate(MaxHook = ifelse(FishingType %in% "Trolling", 0, MaxHook)) %>% 
  mutate_each_(funs(factor(.)), c("FishingType", "WhyChoose", "BaitLure", "Site")) %>% 
  mutate(Comments = paste(exComments, nexComments, aComments, sep = "|"))
```

### Catch data
```{r Catch data}
dat <- dat %>%
  mutate(CaughtUndam = as.numeric(RelUndam + KeptUndam)) %>% 
  mutate(nHooked = as.numeric(CaughtUndam + nDP)) %>% 
  mutate(binDP = ifelse(nDP > 0, "1","0")) %>% 
  mutate(binDP = as.factor(binDP)) %>% 
  mutate(perDP = ifelse(nHooked %in% "0", 0, as.numeric(round((nDP/nHooked)*100, 2)))) %>% 
  mutate(proDP = as.numeric(round(perDP/100, 2))) %>% 
  mutate(nUndamHr = as.numeric(round(CaughtUndam/decDuration, 2))) %>% 
  mutate(nDPHr = as.numeric(round(nDP/decDuration, 2))) %>% 
  mutate(nHookedHr = as.numeric(round(nHooked/decDuration, 2))) %>% 
  mutate(DPrateHr = as.numeric(round(nDPHr/nHookedHr, 2))) %>% 
  mutate(DPrateHr = ifelse(is.nan(DPrateHr), 0, DPrateHr))
```

### Species Caught
```{r Species}
# 
# # Species Caught
# # Mostly done, just some issues with general fish species e.g. bream - could actually be a bream but could also be a common name for a snapper or a emperor
# Ning <- Ning %>%
#   mutate(Emperors = ifelse(str_detect(Species, "blue lined|blue line|black snapper|spangled|Spangled|spango|Spango|Spangley|spangley|spangly|Grassy|red spot|Long nose|Long tongs|REd emp|Red emperor|Red lipped
#                                       |Blue lined emp|Blue lines emp|Emperors|Emporer|slang emperor"), 1, NA)) %>%
#   mutate(Cods_Trouts_Groupers = ifelse(str_detect(Species, "estuarine cod|Charlie|charlie|rock cod|rockcod|sweetlips|coral|Coral|honeycomb|Honeycomb|honey|Honey|rankin|Coronation|coronation|nanny|Chinaman|chinamen|chinaman
#                                                   |cod|Cod|Cud’s"), 1, NA)) %>%
#   mutate(Snappers = ifelse(str_detect(Species, "Rims on|Red throat|Flag|flags|Ruby|Goldband|goldband|Goldban|Gold band||Golden band|gold and|Mangrove|mangrove|Jack|jack|Red snapper|Spainish|Striped sea|stripey sera"), 1, NA)) %>%
#   mutate(Breams = ifelse(str_detect(Species, "Robinson's|seabream|Seabream|Pink bream|thredfin|threadfin"), 1, NA)) %>%
#   mutate(Perch = ifelse(str_detect(Species, "Pearl perch|Dhufish|dhufosh"), 1, NA)) %>%
#   mutate(Pelagic = ifelse(str_detect(Species, "black marlin|sailfish|Marlin|barracude|trevally|Trevally|trivali|trevallu|Cobia|Kobe|Obie|queenie|Queenie|queen|tuna|Yellowfin|mackerel|Mackerel|mackrel|Mackeral|mac,real|Mackey
#                                      |Mahi|bludger|Wahoo|Waco"), 1, NA)) %>%
#   mutate(Flatheads_Cobblers = ifelse(str_detect(Species, "Cobbler|Flathead|Flatty|catfish|Northern flathead|Sandbar flathead"), 1, NA)) %>%
#   mutate(Pigfish_Tuskfish_Parrotfish = ifelse(str_detect(Species, "blue bone|black spot|parrot|vine"), 1, NA)) %>%
#   mutate(Cephalopods_Crustaceans = ifelse(str_detect(Species, "Squid|squid|Squids|squids|Prawn|prawn|Prawns|prawns"), 1, NA)) %>%
#   mutate(Others = ifelse(str_detect(Species, "Blowfish|Blowly|northwest|Damsel|Leather jacket|Lizard fish|Shark|Whalers"), 1, NA)) %>%
#   relocate(Emperors, .after=Species) %>%
#   relocate(Snappers, .after=Emperors) %>%
#   relocate(Breams, .after=Snappers) %>%
#   relocate(Cods_Trouts_Groupers, .after=Breams) %>%
#   relocate(Perch, .after=Cods_Trouts_Groupers) %>%
#   relocate(Pelagic, .after=Perch) %>%
#   relocate(Flatheads_Cobblers, .after=Pelagic) %>%
#   relocate(Pigfish_Tuskfish_Parrotfish, .after=Flatheads_Cobblers) %>%
#   relocate(Cephalopods_Crustaceans, .after=Pigfish_Tuskfish_Parrotfish) %>%
#   relocate(Others, .after=Cephalopods_Crustaceans)
# 
# unique(dat$Species)
```


## Avidity and demographics cleaning
```{r avidity clean}
dat <- dat %>% 
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 0, 4), "0-4", NA)) %>% 
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 5, 9), "5-9", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 10, 14), "10-14", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 15, 19), "15-19", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 20, 29), "20-29", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 30, 39), "30-39", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(between(exTimes12m, 40, 59), "40-59", bin_exTimes12m)) %>%
  mutate(bin_exTimes12m = ifelse(exTimes12m > 59, "60+", bin_exTimes12m)) 
# Age 
dat <- dat %>% 
  mutate(Age = 2021 - YrBorn) %>% 
  mutate(binAge = ifelse(between(Age, 18, 25), "18-24", NA)) %>% 
  mutate(binAge = ifelse(between(Age, 25, 34), "25-34", binAge)) %>%
  mutate(binAge = ifelse(between(Age, 35, 44), "35-44", binAge)) %>%
  mutate(binAge = ifelse(between(Age, 45, 54), "45-54", binAge)) %>%
  mutate(binAge = ifelse(between(Age, 55, 64), "55-64", binAge)) %>%
  mutate(binAge = ifelse(between(Age, 65, 74), "65-74", binAge)) %>%
  mutate(binAge = ifelse(Age > 74, "75+", binAge))

# Postcode
# dat <- dat %>% 
#   mutate(Postcode = as.factor(ifelse(Resident == "Yes", "6450", Postcode))) %>% 
#   mutate(Postcode = as.factor(ifelse(Postcode == " 0", NA, Postcode))) 

# Party
dat <- dat %>% 
  mutate(Sex = as.factor(Sex)) %>% 
  mutate(nFemales = ifelse(is.na(nFemales), 0, nFemales)) %>%
  mutate(nMales = ifelse(is.na(nMales), 0, nMales)) %>%
  mutate(nGirls = ifelse(is.na(nGirls), 0, nGirls)) %>%
  mutate(nBoys = ifelse(is.na(nBoys), 0, nBoys)) %>%
  mutate(nFemales = ifelse(Sex %in% "Female", nFemales + 1, nFemales)) %>%
  mutate(nMales = ifelse(Sex %in% "Male", nMales + 1, nMales)) %>% 
  mutate(nAdults = nFemales + nMales) %>% 
  mutate(nChild = nGirls + nBoys) %>% 
  mutate(PartySize = nAdults + nChild) %>% 
  mutate_each_(funs(factor(.)), c("Accom", "DiveCert")) %>% 
  mutate(kids = ifelse(nChild > 0, 1, 0))
```

## Lunar Phase
```{r lunar phase}
dat <- dat %>%
  mutate(LunarPhase = as.factor(lunar.phase(TripDate, name = TRUE)))
```

```{r}
dat <- dat %>% 
  dplyr::select(
    ID, TripID, SurveyDate, numYear, facYear, Interviewer, mLat, mLong, PersonID, Screen18,
         PrevInter, Agreement, binRecall, TripDate, TripMonth, TripJulianDay, TripNum, SiteNum,
         Site, SiteType, Resident, BoatAccess, nDaysInArea, nBoatDays, nShoreDays, nTimesLast12m,
         UseLat, UseLong, ActivityType, Activity, Start, Stop, MedianTime, decMedianTime,
         decDuration, FishingType, BaitLure, MaxHook, KeptUndam, RelUndam, nDP, CaughtUndam,
         nHooked, binDP, perDP, proDP, CaughtUndam, nHooked, nUndamHr, nDPHr, DPrateHr, Species,
         DiveMaxDepth, WhyChoose, WhyLeave, Detract, exTimes12m, bin_exTimes12m, exYrs, nexTimes12m, nexYrs,
         DiveCert, nDives, Age, binAge, Postcode, Accom, Sex, nFemales, nMales, nGirls, nBoys, nAdults,
         nChild, kids, PartySize, BoatLength, BoatID, Comments, LunarPhase)
```

# Check point
```{r csv}
write.csv(dat, "./data/01_data/1.1_dat.csv", row.names = FALSE) # Fully cleaned data
```
