==> 2.0_StaticCov_5km.Rmd <==

```{r setup}
# knitr options
knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = FALSE, fig.align = 'center', fig.width = 10, fig.height = 7) 

# libraries 
library(tidyverse)
library(dplyr)
library(ggplot2)
library(sp)
library(raster)
library(rgeos)
library(rgdal)
library(sf)
library(googledrive)
library(units)
library(nngeo)

# functions
source("./functions/spatialFunc.R") # spatial functions

# data
## spatial
### gpkg - ESPG 4283 (GDA94)
BR <- st_read("./data/gpkg/BR_4283.gpkg") %>%  st_transform(4283)# boat ramps
coast <- st_read("./data/gpkg/Coast_4283.gpkg") %>% st_transform(4283)# coast
# habitat <- st_read("./data/gpkg/espHabitat_4283.gpkg") # habitat
sites <- st_read("./data/gpkg/mar5kmgrid.gpkg") %>%  st_transform(4283)# MVM grid
studySite <- st_read("./data/gpkg/StudySite_4283.gpkg") %>% st_transform(4283) # study site (mega mvm polygon)
# artInfra <- st_read("./data/gpkg/espArtInfra_4283.gpkg")
shelter <- st_read("./data/gpkg/mar5km_shelter.gpkg") %>% st_transform(4283)
NPZ <- st_read("./data/gpkg/ntz_4283.gpkg") %>% st_transform(4283)# commonwealth and current state NTZ

### raster - ESPG 4326 (WGS84)
bathy <- raster("./data/raster/bathy_4326.tif") # bathy

sf_use_s2(FALSE)

# make sure all layers overlap
# ggplot()+
  # geom_sf(data = sites, lwd = 0.05) +
  # geom_sf(data = shelter, fill = 'red', colour = 'red') +
  # geom_sf(data = BR) +
  # geom_sf(data = NPZ)
```

# Static Variables
This script add static spatial variables to all cells in a grid. 

    - unique ID
    - mean depth
    - centroid of grid
    - area of every habitat
    - area of every grid
    - presence of adjacent island
    - presence of artificial infrastructure (eg. FADS, artificial reefs)
    - distance to boat ramp
    - distance to mainland
    - southern fetch (shelter)
    - commonwealth NPZ

# shelter 
Needs to be done before you manipulate IDs
```{r}
sites$shelter <- ifelse(sites$fid %in% shelter$fid, 1, 0)

ggplot() +
  geom_sf(data = sites, aes(fill = shelter), lwd = 0.05)
```

## Assign unique identifier to each grid cell
For Esperance is needs to match the mapper and add in ids for new cells. 
```{r gridID}
sites <- sites %>% 
  dplyr::select(geom) %>% 
  mutate(gridID = row_number()) %>% 
  st_make_valid()
```

## Depth
Calculates the average depth within each grid cell (site). This does work is after grid manipulation.  

```{r depth}
sites <- mean_depth(bathy, sites) # get average depth of grid

ggplot() +
  geom_sf(data = sites, aes(fill = Depth), lwd = 0.05)
```

## Data prep
  - waterGrid: sf (polygon) of the grid which is only on the water
  - polyCentroid: sf (point) which is the centroid of each grid cell/site if water_grid
  
```{r data prep}
sf_use_s2(FALSE)

waterGrid <- st_difference(sites, st_combine(coast)) # isolates grid to water
polyCentroid <- st_centroid_within_poly(waterGrid) # gets centroid of every grid cell
waterGrid$polyCentroid <- polyCentroid$geom # bind centroid to grid sf 

ggplot() +
  geom_sf(data = waterGrid$polyCentroid)
```

## Habitat
Calculates the area (km2) of each habitat type in each individual grid cell (site)
```{r habitat}
# waterGrid <- habitat_area(sf_habitat = habitat, col_habitat=1, sf_grid = waterGrid, id_grid = ID)
```

## Area
Calculates the area of each polygon
```{r area}
waterGrid$area <- round(set_units(st_area(waterGrid), km^2), 2) 

# ggplot() +
#   geom_sf(data = waterGrid, aes(fill = as.numeric(area)))

waterGrid <- waterGrid %>%
  mutate(hex = ifelse(as.numeric(area) >= 0.6, "large", NA)) %>%
  mutate(hex = ifelse(as.numeric(area) >= 0.1 & as.numeric(area) <= 0.6, "medium", hex)) %>%
  mutate(hex = ifelse(as.numeric(area) <= 0.1, "small", hex)) %>%
  mutate(hex = as.factor(hex)) %>%
  mutate(hex_num = ifelse(hex %in% "large", 1, NA)) %>%
  mutate(hex_num = ifelse(hex %in% "medium", 2, hex_num)) %>%
  mutate(hex_num = ifelse(hex %in% "small", 3, hex_num)) %>%
  mutate(hex_num = as.numeric(hex_num))

waterGrid <- waterGrid %>% 
  mutate(Lhex = ifelse(as.numeric(area) >= 0.6, 1, 0)) %>% 
  mutate(Mhex = ifelse(as.numeric(area) > 0.1 & as.numeric(area) < 0.6, 1, 0)) %>% 
  mutate(Shex = ifelse(as.numeric(area) <= 0.1, 1, 0))

waterGrid <- waterGrid %>% 
  mutate(commhex = ifelse(as.numeric(area) >= 22, 1, 0))

# dat %>%
#   group_by(hex) %>%
#   summarise(sum(choice))
#   Not enough data at small spatial scales for model to work. add this as a dummy variable instead of area.
#   
ggplot() +
  geom_sf(data = waterGrid, aes(fill = commhex))
```

## Artificial infrastructure present
Creates binary list indicating whether there is artificial infrastructure present in grid cell (site)

```{r artificial infrastructure present}
# waterGrid$ArtInfra_present <- as.numeric(st_intersects(artInfra, waterGrid, sparse = FALSE)) 
```

## Land
Distance to mainland: Calculates the linear distance from the centroid of every grid cell (site) to the nearest point on mainland.
Adjacent to land: Creates binary list indicating whether there is an island are adjacent to grid cell (site).
```{r land present}
coast$area <- as.numeric(set_units(st_area(coast), km^2)) # area of polygons in coast sf

# Distance from mainland
## extracting mainland from coast polygon
mainland <- coast %>% 
  arrange(desc(area)) %>% 
  slice(1)

# ## calculate distance form mainland
km_mainland <- dist_mainland(waterGrid$polyCentroid, mainland)
# 
waterGrid$km_mainland <- km_mainland[,1]

# Adjacent to land
## extracting islands from coast polygon
islands <- coast %>%
  arrange(desc(area))  %>%
  slice(2:nrow(coast))

## create binary column indicating whether grid cell is adjacent to island
waterGrid %<>%
  mutate(isl_adj = ifelse(as.list(st_intersects(waterGrid, islands)) %in% "integer(0)", 0, 1))

ggplot() +
  geom_sf(data = waterGrid, aes(fill = km_mainland))

ggplot()+
  geom_sf(data = waterGrid, aes(fill = isl_adj))
```

# Fetch
Calculating the Southerly and Easterly fetch from the centroid of all polygons on every day. If fetch is over 2.5km then 1 else 0.
```{r fetch}
# coast_lambert <- coast %>% st_transform(crs = 3112) # need land lambert
# 
# cent <- waterGrid %>%  st_drop_geometry() %>% st_as_sf() # dropping geom
# 
# fetch <- windfetch(polygon_layer = coast_lambert, site_layer = cent) # calc fetch
```

## Shelter
Fetch was too computationally heavy for my time scale, so i've made a shape file with all the grids that have shelter present.
```{r}
# shelter <- shelter %>% st_transform(crs(coast))
# waterShelter <- st_difference(shelter, st_combine(coast)) # isolate
# 
# ggplot() +
#   geom_sf(data = waterShelter)

# shelter_grids <- as.data.frame(st_within(waterGrid, waterShelter, sparse = FALSE))
# 
# shelter_grids <- shelter_grids %>%
#   rowwise %>%
#   mutate(shelter = if_any(.cols = contains('V'), isTRUE)) %>%
#   ungroup
# 
# waterGrid <- waterGrid %>%
#   mutate(shelter = ifelse(shelter_grids$shelter %in% TRUE, 1, 0))

# ggplot() +
#   geom_sf(data = waterGrid, aes(fill = shelter), lwd = 0.05)
```

## Commonwealth NPZ
Adding a binary variable as to whether a grid is within a NPZ, if so it will need to have 0 utility for recreational fishers. 
```{r}
# npz <- st_union(NPZ)
# 
# npz_grids <- as.list(st_intersects(waterGrid, npz, sparse = FALSE))
# 
# waterGrid <- waterGrid %>% 
#   mutate(NPZ = ifelse(npz_grids == TRUE, 1, 0))
# 
# ggplot() +
#   geom_sf(data = waterGrid, aes(fill = NPZ), lwd = 0.05)
```

# Distance to BR
Calculates the linear distance from each boat ramp to every centroid - data frame rows should multiple by number of boat ramps.

```{r distance to BR}
dist <- as.data.frame(set_units(st_distance(waterGrid$polyCentroid, BR), km)) 
names(dist) <- unique(BR$RampID)
gridID <- waterGrid$gridID
dist$gridID <- gridID
dist <- dist %>% gather("BR", "km_BR", 1:length(dist) - 1) 
waterGrid <- waterGrid %>% left_join(dist, by = "gridID") # join to input data
```

## Save output
```{r save}
saveRDS(waterGrid, "./data/gpkg/2.0_5kmSites_4283.gpkg")
```
